requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
cloudinary
django-cloudinary-storage
mysqlclient
mercadopago
django-jazzmin
